1120428-0036

+++ app/Http/Controllers/UserController.php
@@ -150,9 +150,11 @@ class UserController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(User $user)
+    public function destroy(User $user, $is_delete)
     {
-        $user->delete();
+        if ($is_delete == 'Y') {
+            $user->delete();
+        }
         return redirect(route('users.index'));
     }
